integrate VNPay

+++ shirts/views.py
@@ -18,6 +18,8 @@ from django.contrib.auth.models import User
 from paypal.standard.forms import PayPalPaymentsForm
 from django.views.decorators.csrf import csrf_exempt
 from store.settings import PAYPAL_RECEIVER_EMAIL
+from shirts.vnpay import vnpay
+from datetime import datetime
 
 
 
@@ -381,15 +383,42 @@ def checkout(request):
                 form = PayPalPaymentsForm(initial=paypal_dict)
                 return render(request, 'process_payment.html', {'order': orders, 'form': form})
 
-
             else:
-                finalorders = orders
-                finalorders.order_status = 'PLACED'
-                finalorders.save()
-                cart = []
-                request.session['cart'] = cart
-                Cart.objects.filter(user=user).delete()
-                return redirect('orders')
+                if payment_method == 'VNPAY':
+                    if True:
+                        ipaddr = get_client_ip(request)
+                        # Build URL Payment
+                        vnp = vnpay()
+                        vnp.requestData['vnp_Version'] = '2.1.0'
+                        vnp.requestData['vnp_Command'] = 'pay'
+                        vnp.requestData['vnp_TmnCode'] = 'XBQV9IFW'
+                        vnp.requestData['vnp_Amount'] = 100 * 100000
+                        vnp.requestData['vnp_CurrCode'] = 'VND'
+                        vnp.requestData['vnp_TxnRef'] =  orders.id
+                        vnp.requestData['vnp_OrderInfo'] = 'Thanh toan don hang o HCMUE Ecomerce'
+                        vnp.requestData['vnp_OrderType'] = 'abc'
+                        # Check language, default: vn
+                        if True:
+                            vnp.requestData['vnp_Locale'] = 'vn'
+                        else:
+                            vnp.requestData['vnp_Locale'] = 'vn'
+                        vnp.requestData['vnp_CreateDate'] = datetime.now().strftime('%Y%m%d%H%M%S')  # 20150410063022
+                        vnp.requestData['vnp_IpAddr'] = ipaddr
+                        vnp.requestData['vnp_ReturnUrl'] = 'http://127.0.0.1:8000/orders/'
+                        vnpay_payment_url = vnp.get_payment_url('https://sandbox.vnpayment.vn/paymentv2/vpcpay.html', 'WVLQMDWIHYIWUMHVQEGOJKOXEWHXQGZH')
+                        print(vnpay_payment_url)
+                        return redirect(vnpay_payment_url)
+                    else:
+                        return render(request, 'payment_failed.html')
+
+                else:
+                    finalorders = orders
+                    finalorders.order_status = 'PLACED'
+                    finalorders.save()
+                    cart = []
+                    request.session['cart'] = cart
+                    Cart.objects.filter(user=user).delete()
+                    return redirect('orders')
         else:
             print(">>>form not valid")
             return redirect('checkout')
@@ -559,3 +588,12 @@ def profile(request):
     users = user = get_object_or_404(User, id=request.user.id)
 
     return render(request, 'profile.html', {'users': users})
+
+
+def get_client_ip(request):
+    x_forwarded_for = request.META.get('HTTP_X_FORWARDED_FOR')
+    if x_forwarded_for:
+        ip = x_forwarded_for.split(',')[0]
+    else:
+        ip = request.META.get('REMOTE_ADDR')
+    return ip
